More logic on search rooms for a selected period of time! Now we can add  extra features for our reservation
Short summary:
Now you can search for available rooms for a certain selected period of time, when you click on search your rooms appear, here you have to select a room, then depending on the selected room you will see the available extra options that you can add to your future booking. also, after selecting the extra options and pressing the total price button of the extra options per day, it shows us the price. Now we are going to make the booking logic based on what has been selected so far and the creation of the invoices